+++ src/main/java/com/rapidtech/restapi/repository/CustomerRepo.java
@@ -3,5 +3,5 @@ package com.rapidtech.restapi.repository;
 import com.rapidtech.restapi.entity.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepo extends JpaRepository<CustomerEntity, Long> {
 }
